move testAccPreCheck to provider_test
