+++ provider/resource_elkaliases_index_test.go
@@ -3,7 +3,6 @@ package provider
 import (
 	"encoding/json"
 	"fmt"
-	"os"
 	"regexp"
 	"testing"
 
@@ -484,15 +483,6 @@ func TestAccElkaliasesIndex_invalid(t *testing.T) {
 			},
 		})
 	})
-}
-
-func testAccPreCheck(t *testing.T) {
-	if v := os.Getenv("ELASTICSEARCH_ENDPOINT"); v == "" {
-		t.Fatal("ELKALIASES_URL must be set for acceptance tests")
-	}
-	if v := os.Getenv("ELASTICSEARCH_API_KEY"); v == "" {
-		t.Fatal("ELKALIASES_TOKEN must be set for acceptance tests")
-	}
 }
 
 func getIndexTemplate(id string) (map[string]any, error) {
